[doc] Add docs for metric sequence sampling
* Move all data update/rollback commands under `storage`
