+++ aim/cli/configs.py
@@ -1,8 +1,8 @@
 INIT_NAME = 'init'
 VERSION_NAME = 'version'
 UP_NAME = 'up'
-UPGRADE_NAME = 'upgrade'
 REINDEX_NAME = 'reindex'
 SERVER_NAME = 'server'
 RUNS_NAME = 'runs'
 CONVERT = 'convert'
+STORAGE = 'storage'
